Renaming Inspalation to InstalationPage. Refactoring
