+++ README.md
@@ -51,3 +51,15 @@ Use next information for prompts during setup:
 npm run dev
 ```
 2. Open `http://localhost:8080/`
+
+
+## Deployment Instructions for Git Branch
+## To deploy your Git branch to AWS S3, please follow these steps:
+
+1. Navigate to the "Deploy to AWS S3" action in your repository.
+2. Choose "Run Workflow" from the dropdown menu.
+3. Enter the name of your branch and click on the "Run workflow" button.
+4. Wait for the deployment process to complete. You can check the progress in the "Actions" tab of your repository.
+5. Once the deployment is finished, verify that your changes have been deployed by checking the website at the following URL: https://reportportal.github.io/.
+
+That's it! Your changes should now be live on the website. If you encounter any issues during the deployment process, please consult the documentation or reach out to the project maintainers for assistance.
