+++ src/components/InstallationPage/InstallationPage.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useBoolean } from 'ahooks';
 
-import { Container } from '../Container';
 import { createBemBlockBuilder } from '../../utils';
 import { DockerContent } from './components/DockerContent';
 import { KubernetesContent } from './components/KubernetesContent';
@@ -9,7 +8,7 @@ import { LaunchPortal } from './components/LaunchPortal';
 import { DockerIcon, KubernetesIcon } from './icons';
 import { SwitchButtons } from '../SwitchButtons';
 
-import './Installation.scss';
+import './InstallationPage.scss';
 
 const getBlocksWith = createBemBlockBuilder(['installation']);
 
@@ -26,8 +25,8 @@ const buttons = [
   },
 ];
 
-export const Installation = () => {
-  const [contentSate, { toggle }] = useBoolean(true);
+export const InstallationPage = () => {
+  const [contentState, { toggle }] = useBoolean(true);
 
   const handleSwitch = () => {
     toggle();
@@ -36,17 +35,17 @@ export const Installation = () => {
   return (
     <div>
       <div className={getBlocksWith()}>
-        <Container>
+        <div className={getBlocksWith('__container')}>
           <h1 className={getBlocksWith('__title')}>Installation guide</h1>
           <p className={getBlocksWith('__subtitle')}>3 steps to get started with ReportPortal</p>
 
           <div className={getBlocksWith('__btn-box')}>
             <SwitchButtons buttons={buttons} onSwitch={handleSwitch} />
           </div>
-        </Container>
+        </div>
       </div>
 
-      {contentSate ? <DockerContent /> : <KubernetesContent />}
+      {contentState ? <DockerContent /> : <KubernetesContent />}
 
       <LaunchPortal />
     </div>
