+++ src/components/InstallationPage/InstallationPage.scss
@@ -3,28 +3,20 @@
 @import 'src/styles/variables';
 @import 'src/styles/mixins/breakpoint';
 
-.code-text {
-  display: inline-block;
-  margin: 5px 0;
-  padding: 2px 10px;
-  border: 1px solid var(--bg-purple-grey);
-  border-radius: 8px;
-  color: var(--text-service);
-}
-
-.sub-text {
-  color: var(--text-light-grey);
-}
-
 .installation {
   position: relative;
-  padding: 72px 16px 48px;
-  margin-top: $header-height;
+  padding: calc(72px + $header-height) 16px 48px;
+  margin-top: -$header-height;
   color: var(--text-white);
   background: var(--color-primary-600);
   background-image: url('./icons/group-244.svg');
   background-repeat: no-repeat;
   background-size: cover;
+
+  &__container {
+    padding: 0 20px;
+    margin: 0 auto;
+  }
 
   &__title {
     @include font-poppins($fw-bold);
@@ -198,4 +190,17 @@
       flex-basis: 48.5%;
     }
   }
+}
+
+.code-text {
+  display: inline-block;
+  margin: 5px 0;
+  padding: 2px 10px;
+  border: 1px solid var(--bg-purple-grey);
+  border-radius: 8px;
+  color: var(--text-service);
+}
+
+.sub-text {
+  color: var(--text-light-grey);
 }
