+++ src/components/InstallationPage/components/DockerContent/index.js
