+++ src/components/InstallationPage/components/KubernetesContent/index.js
