+++ src/components/InstallationPage/components/Notice/Notice.scss
