+++ src/components/InstallationPage/icons/DockerIcon.jsx
