+++ src/components/InstallationPage/icons/KubernetesIcon.jsx
