+++ src/components/InstallationPage/icons/index.js
