+++ src/pages/installation.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Layout } from '../components/Layout';
-import { Installation as InstallationPage } from '../components/Installation';
+import { InstallationPage } from '../components/InstallationPage';
 
 import '../styles/global.scss';
