+++ static/github.json
@@ -1,1 +1,1 @@
-{"total":2574,"repos":{".github":0,"JBehaveSerenityAutomation":0,"JavaSDK":0,"agent-Python-RobotFramework":49,"agent-android":1,"agent-java-cucumber":22,"agent-java-cucumber2":0,"agent-java-cucumber3":0,"agent-java-cucumber4":7,"agent-java-cucumber5":1,"agent-java-cucumber6":6,"agent-java-jbehave":7,"agent-java-junit":10,"agent-java-junit5":20,"agent-java-karate":9,"agent-java-soapui":4,"agent-java-spock":8,"agent-java-test-utils":0,"agent-java-testNG":46,"agent-js-codecept":8,"agent-js-cucumber":8,"agent-js-cypress":35,"agent-js-jasmine":9,"agent-js-jest":21,"agent-js-mocha":15,"agent-js-nightwatch":2,"agent-js-playwright":19,"agent-js-postman":16,"agent-js-testcafe":5,"agent-js-webdriverio":3,"agent-net-gauge":10,"agent-net-nunit":19,"agent-net-nunit2":2,"agent-net-specflow":8,"agent-net-vstest":8,"agent-net-xunit":4,"agent-php-Behat":0,"agent-php-Codeception":1,"agent-php-PHPUnit":0,"agent-python-behave":6,"agent-python-nosetests":0,"agent-python-pytest":86,"agent-readyapi":0,"agent-ruby":15,"agent-scala-scalatest":2,"agent-silktest":0,"agent-swift-XCTest":4,"aggregator-chart":0,"allure-java":6,"android-java-example":1,"android-kotlin-example":1,"ansible-playbook":1,"client-Python":36,"client-java":19,"client-java-core":6,"client-java-httpclient-repacked":3,"client-java-rest-core":3,"client-javascript":15,"client-net":13,"client-php":0,"commons":6,"commons-bom":0,"commons-bugtracking":4,"commons-codec-repacked":5,"commons-dao":12,"commons-events":0,"commons-fonts":0,"commons-go":2,"commons-model":8,"commons-net":5,"commons-rules":5,"complex-migrations":0,"data-migration":0,"docs":25,"environment-wandereryellow-production":0,"environment-wandereryellow-staging":0,"es-logs-service":0,"example-cucumber-junit-selenium-logback-maven":2,"example-custom-analyzer":4,"example-groovy-spock-logback":1,"example-net-gauge":1,"example-net-nunit":11,"example-net-nunit2":1,"example-net-specflow":4,"example-net-specflow-nunit2-log4net":1,"example-net-vstest":1,"example-net-xunit":0,"example-service-java":0,"example-spock-webdriver-logback-maven":0,"example-testng-selenium-log4j-gradle":0,"examples-java":14,"examples-javascript":0,"examples-js":6,"examples-python":5,"examples-scala":0,"goRP":0,"gradle-scripts":1,"jbehave-tutorial":0,"jira-versions-sync":0,"kubernetes":68,"landing-aggregator":3,"linux-installation":3,"logger-java-httpcomponents":0,"logger-java-log4j":13,"logger-java-logback":6,"logger-java-okhttp3":0,"logger-java-rest-assured":1,"logger-java-selenide":1,"logger-net-log4net":1,"logger-net-nlog":1,"logger-net-serilog":1,"logger-net-tracelistener":1,"migrations":1,"nestjs-rest-api":0,"plugin-api":2,"plugin-auth-saml":0,"plugin-bts-azure":2,"plugin-bts-jira":9,"plugin-bts-jira-cloud":0,"plugin-bts-rally":2,"plugin-example":1,"plugin-jenkins":4,"plugin-saucelabs":4,"plugin-template":0,"reportportal":1548,"reportportal-common-api":0,"reportportal-maven-plugin":0,"reportportal.github.io":11,"restendpoint":1,"rs-cart-api":0,"service-analyzer":5,"service-analyzer-equals":1,"service-api":37,"service-authorization":16,"service-auto-analyzer":8,"service-gateway":9,"service-index":3,"service-jira":9,"service-jobs":0,"service-metrics-gatherer":0,"service-rally":5,"service-registry":8,"service-tfs":6,"service-ui":53,"shell-installation":4,"ui-kit":0,"utils-java-formatting":2,"versions":0}}
+{"total":2575,"repos":{".github":0,"JBehaveSerenityAutomation":0,"JavaSDK":0,"agent-Python-RobotFramework":49,"agent-android":1,"agent-java-cucumber":22,"agent-java-cucumber2":0,"agent-java-cucumber3":0,"agent-java-cucumber4":7,"agent-java-cucumber5":1,"agent-java-cucumber6":6,"agent-java-jbehave":7,"agent-java-junit":10,"agent-java-junit5":20,"agent-java-karate":9,"agent-java-soapui":4,"agent-java-spock":8,"agent-java-test-utils":0,"agent-java-testNG":46,"agent-js-codecept":8,"agent-js-cucumber":8,"agent-js-cypress":35,"agent-js-jasmine":9,"agent-js-jest":21,"agent-js-mocha":15,"agent-js-nightwatch":2,"agent-js-playwright":19,"agent-js-postman":16,"agent-js-testcafe":5,"agent-js-webdriverio":3,"agent-net-gauge":10,"agent-net-nunit":19,"agent-net-nunit2":2,"agent-net-specflow":8,"agent-net-vstest":8,"agent-net-xunit":4,"agent-php-Behat":0,"agent-php-Codeception":1,"agent-php-PHPUnit":0,"agent-python-behave":6,"agent-python-nosetests":0,"agent-python-pytest":86,"agent-readyapi":0,"agent-ruby":15,"agent-scala-scalatest":2,"agent-silktest":0,"agent-swift-XCTest":4,"aggregator-chart":0,"allure-java":6,"android-java-example":1,"android-kotlin-example":1,"ansible-playbook":1,"client-Python":37,"client-java":19,"client-java-core":6,"client-java-httpclient-repacked":3,"client-java-rest-core":3,"client-javascript":15,"client-net":13,"client-php":0,"commons":6,"commons-bom":0,"commons-bugtracking":4,"commons-codec-repacked":5,"commons-dao":12,"commons-events":0,"commons-fonts":0,"commons-go":2,"commons-model":8,"commons-net":5,"commons-rules":5,"complex-migrations":0,"data-migration":0,"docs":25,"environment-wandereryellow-production":0,"environment-wandereryellow-staging":0,"es-logs-service":0,"example-cucumber-junit-selenium-logback-maven":2,"example-custom-analyzer":4,"example-groovy-spock-logback":1,"example-net-gauge":1,"example-net-nunit":11,"example-net-nunit2":1,"example-net-specflow":4,"example-net-specflow-nunit2-log4net":1,"example-net-vstest":1,"example-net-xunit":0,"example-service-java":0,"example-spock-webdriver-logback-maven":0,"example-testng-selenium-log4j-gradle":0,"examples-java":14,"examples-javascript":0,"examples-js":6,"examples-python":5,"examples-scala":0,"goRP":0,"gradle-scripts":1,"jbehave-tutorial":0,"jira-versions-sync":0,"kubernetes":68,"landing-aggregator":3,"linux-installation":3,"logger-java-httpcomponents":0,"logger-java-log4j":13,"logger-java-logback":6,"logger-java-okhttp3":0,"logger-java-rest-assured":1,"logger-java-selenide":1,"logger-net-log4net":1,"logger-net-nlog":1,"logger-net-serilog":1,"logger-net-tracelistener":1,"migrations":1,"nestjs-rest-api":0,"plugin-api":2,"plugin-auth-saml":0,"plugin-bts-azure":2,"plugin-bts-jira":9,"plugin-bts-jira-cloud":0,"plugin-bts-rally":2,"plugin-example":1,"plugin-jenkins":4,"plugin-saucelabs":4,"plugin-template":0,"reportportal":1548,"reportportal-common-api":0,"reportportal-maven-plugin":0,"reportportal.github.io":11,"restendpoint":1,"rs-cart-api":0,"service-analyzer":5,"service-analyzer-equals":1,"service-api":37,"service-authorization":16,"service-auto-analyzer":8,"service-gateway":9,"service-index":3,"service-jira":9,"service-jobs":0,"service-metrics-gatherer":0,"service-rally":5,"service-registry":8,"service-tfs":6,"service-ui":53,"shell-installation":4,"ui-kit":0,"utils-java-formatting":2,"versions":0}}
